+++ release-notes.rst
@@ -2,6 +2,15 @@ Release notes for chevah.compat
 ===============================
 
 
+0.44.0 - 01/08/2017
+-------------------
+
+* Remove port listening helpers.
+* Update to latest Solaris on 32bit.
+* Add support for OS detection in test case and no longer use hostname
+  to detect the OS.
+
+
 0.43.3 - 08/05/2017
 -------------------